chore: move context closer to challenges
