+++ apps/web/src/app/challenge/[slug]/challenge-route-data.hook.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { createContext, useState, useContext, type PropsWithChildren } from 'react';
-import type { GetCurrentChallengeType } from './challenge/[slug]/getChallengeRouteData';
+import type { GetCurrentChallengeType } from './getChallengeRouteData';
 
 interface ChallengeRouteDataContextType {
   currentChallenge: GetCurrentChallengeType | null;
