+++ apps/web/src/components/Navigation/problem-explorer-track-nav.tsx
@@ -1,5 +1,4 @@
 'use client';
-import { useChallengeRouteData } from '~/app/challenge-route-data.hook';
 import { useProblemExplorerContext } from '~/app/problem-explorer.hooks';
 import { useMemo } from 'react';
 import { ExploreDrawer } from '~/app/challenge/_components/explore-drawer';
@@ -8,6 +7,7 @@ import Link from 'next/link';
 import { ChevronLeft, ChevronRight, Swords } from '@repo/ui/icons';
 import { cn } from '@repo/ui/cn';
 import { Button } from '@repo/ui/components/button';
+import { useChallengeRouteData } from '~/app/challenge/[slug]/challenge-route-data.hook';
 
 interface ProblemExplorerTrackNav {
   isCollapsed: boolean;
